feat(server)!: changed code to not throw exceptions
Changed how an empty result is handled, prior to this change exceptions where thrown. Now empty data is returned (null) when no server is found or a request needs a ressource that does not exist. This gives no info on what is wrong on purpos. The code where error codes are generated to give more insight is in the controller.TBD